requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.1.0
MetaTrader5>=5.0.37
joblib>=1.2.0
python-telegram-bot>=20.0
dash>=2.0
plotly>=5.0
tensorflow>=2.10.0